agents/code-writer/pyproject.toml
[project]
name = "meowone-code-writer"
version = "0.1.0"
description = "MeowOne A2A agent — code writing (a2a-sdk)"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk[http-server]>=0.3.25",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
    "uvicorn[standard]>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
